+++ src/nvsnapshots/snapshot_view.py
@@ -4,19 +4,30 @@ Copyright (c) 2025 Peter Triesberger
 For further information see https://github.com/peter88213/nv_snapshots
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
+from datetime import datetime
 from tkinter import ttk
 
 from nvlib.controller.sub_controller import SubController
 from nvlib.gui.widgets.index_card import IndexCard
+from nvlib.novx_globals import STATUS
 from nvsnapshots.nvsnapshots_globals import FEATURE
+from nvsnapshots.nvsnapshots_globals import icons
 from nvsnapshots.nvsnapshots_locale import _
 from nvsnapshots.platform.platform_settings import KEYS
 from nvsnapshots.platform.platform_settings import PLATFORM
 import tkinter as tk
-from nvsnapshots.nvsnapshots_globals import icons
 
 
 class SnapshotView(tk.Toplevel, SubController):
+
+    _COLUMNS = {
+        'id':('ID', 'id_width'),
+        'title':(_('Title'), 'title_width'),
+        'date':(_('Date'), 'date_width'),
+        'words_used':(_('Words'), 'words_used_width'),
+        'words_total':(_('With unused'), 'words_total_width'),
+        'work_phase':(_('Work phase'), 'work_phase_width'),
+    }
 
     def __init__(self, model, view, controller, prefs):
         super().__init__()
@@ -48,6 +59,8 @@ class SnapshotView(tk.Toplevel, SubController):
         #--- Tree for snapshot selection.
         self._treeView = ttk.Treeview(
             self._mainWindow,
+            columns=tuple(self._COLUMNS),
+            show='headings',
             selectmode='browse',
         )
         scrollY = ttk.Scrollbar(
@@ -63,6 +76,17 @@ class SnapshotView(tk.Toplevel, SubController):
             fill='both',
         )
         self._treeView.bind('<<TreeviewSelect>>', self._on_select_node)
+        for colId in self._COLUMNS:
+            colText, colWidth = self._COLUMNS[colId]
+            self._treeView.column(
+                colId,
+                width=int(self.prefs.get(colWidth, 100)),
+            )
+            self._treeView.heading(
+                colId,
+                text=colText,
+                anchor='w',
+            )
 
         #--- "Index card" in the right frame.
         self._indexCard = IndexCard(
@@ -158,12 +182,31 @@ class SnapshotView(tk.Toplevel, SubController):
     def build_tree(self):
         self.reset_tree()
         for snapshotId in self.snapshots:
+            try:
+                displayDate = datetime.fromisoformat(
+                    self.snapshots[snapshotId]['date']
+                ).strftime('%c')
+                status = self.snapshots[snapshotId]['work phase']
+                if status is not None:
+                    workPhase = STATUS[status]
+                else:
+                    workPhase = _('Undefined')
+            except:
+                displayDate = self.snapshots[snapshotId]['date']
+            columns = [
+                snapshotId,
+                self.snapshots[snapshotId]['title'],
+                displayDate,
+                self.snapshots[snapshotId]['words used'],
+                self.snapshots[snapshotId]['words total'],
+                workPhase,
+            ]
             self._treeView.insert(
                 '',
                 'end',
                 snapshotId,
-                text=snapshotId,
-        )
+                values=columns,
+            )
         self._indexCard.bodyBox.config(state='normal')
         self._indexCard.bodyBox.clear()
         self._indexCard.bodyBox.config(state='disabled')
@@ -174,6 +217,9 @@ class SnapshotView(tk.Toplevel, SubController):
     def on_quit(self, event=None):
         self.update_idletasks()
         self.prefs['window_geometry'] = self.winfo_geometry()
+        for i, colId in enumerate(self._COLUMNS):
+            width = self._COLUMNS[colId][1]
+            self.prefs[width] = self._treeView.column(i, 'width')
         self.destroy()
         self.isOpen = False